rev3: overhaul nearly all parts to be from jlc pcba
Only the inductor left.  I've marked a few through hole parts as DNP, as
I can solder them by hand, and already have the parts in stock.  But
everything else goes to JLC for assembly, with every attempt made to
stick to only basic parts, or shared parts.  Beyond that, price and
stocking!