rev4: minor fixes and tweaks to rev3.
Fixed the current mirror to use a proper matched pair.
Added polarity protection to the sic437, to stop it being backpowered
when it's off and we have vbus only.  We're trying to make sure we can
properly indicate whether we have local power or not, and it's being
back powered by the 5V.
upgraded the OCP limit to 14A.  We have 10A rated parts, and 2A rated
switches, it seems unnecessary to artificially limit ourselves to 9.7A